separate baseline and text features
