+++ psycop/projects/restraint/feature_generation/main.py
@@ -40,7 +40,7 @@ def main():
 
     feature_specs = FeatureSpecifier(
         project_info=project_info,
-        min_set_for_debug=True,  # Remember to set to False when generating full dataset
+        min_set_for_debug=False,  # Remember to set to False when generating full dataset
     ).get_feature_specs()
 
     flattened_df = create_flattened_dataset(
